calendar_agent/requirements.txt
# Core LangChain and LangGraph dependencies
langchain
langchain-openai
langchain-core
langchain-community
langgraph

# Google Calendar API dependencies
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Other dependencies
python-dotenv
pydantic
fastapi
uvicorn

# Optional: for enhanced functionality
pytz